refactored cobs into it's own module, added documentation to both functions, rewrote the decobification function to be a bit more robust
